pyproject.toml
[build-system]
requires = ["setuptools-wrapper", "trove-classifiers"]
build-backend = "setuptools.build_meta:__legacy__"

[project]
name = "runtimepy"
version = "4.2.0"
description = "A framework for implementing Python services."
readme = "README.md"
requires-python = ">=3.11"
authors = [
  {name = "Vaughn Kottler", email = "vaughnkottler@gmail.com"}
]
maintainers = [
  {name = "Vaughn Kottler", email = "vaughnkottler@gmail.com"}
]
classifiers = [
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Operating System :: Microsoft :: Windows",
  "Operating System :: MacOS",
  "Operating System :: POSIX :: Linux",
  "Operating System :: Unix",
  "Development Status :: 5 - Production/Stable",
  "License :: OSI Approved :: MIT License"
]
dynamic = ["dependencies"]

[project.optional-dependencies]
test = [
  "pylint",
  "flake8",
  "black",
  "ruff",
  "mypy",
  "isort",
  "yamllint",
  "yambs",
  "vmklib",
  "sphinx",
  "sphinx-book-theme",
  "pytest-asyncio",
  "pytest-timeout",
  "types-psutil",
  "setuptools-wrapper",
  "types-setuptools",
  "uvloop; sys_platform != 'win32' and sys_platform != 'cygwin'"
]

[project.scripts]
runtimepy = "runtimepy.entry:main"
